@@ -18,12 +18,12 @@
 	},
 	"minimum-stability": "dev",
 	"require": {
-		"php": ">=5.3.0",
-		"icanboogie/module-installer": "*",
-		"icanboogie/common": "*"
+		"php": ">=5.5.0",
+		"icanboogie/module-installer": "^1.2.3",
+		"icanboogie/common": "^1.2.2"
 	},
 	"require-dev": {
-		"icanboogie/event": "*"
+		"icanboogie/event": "^1.4.0"
 	},
 	"autoload": {
 		"classmap": [ "lib" ]
