@@ -4,12 +4,14 @@ namespace Icybee\Modules\Editor;
 
 use ICanBoogie\HTTP\Request;
 
-return array
-(
-	'api:editors/new-pane' => array
-	(
+return [
+
+	'api:editors/new-pane' => [
+
 		'pattern' => '/api/editors/tabbable/new-pane',
-		'controller' => __NAMESPACE__ . '\TabbableNewPaneOperation',
+		'controller' => TabbableNewPaneOperation::class,
 		'via' => Request::METHOD_GET
-	)
-);
+
+	]
+
+];
