@@ -19,7 +19,7 @@ class TextEditor implements Editor
 	/**
 	 * Returns content as is.
 	 *
-	 * @see Icybee\Modules\Editor.Editor::serialize()
+	 * @inheritdoc
 	 */
 	public function serialize($content)
 	{
@@ -29,7 +29,7 @@ class TextEditor implements Editor
 	/**
 	 * Returns serialized content as is.
 	 *
-	 * @see Icybee\Modules\Editor.Editor::unserialize()
+	 * @inheritdoc
 	 */
 	public function unserialize($serialized_content)
 	{
@@ -38,7 +38,7 @@ class TextEditor implements Editor
 	/**
 	 * @return TextEditorElement
 	 *
-	 * @see Icybee\Modules\Editor.Editor::from()
+	 * @inheritdoc
 	 */
 	public function from(array $attributes)
 	{
@@ -48,7 +48,7 @@ class TextEditor implements Editor
 	/**
 	 * Returns content as is.
 	 *
-	 * @see Icybee\Modules\Editor.Editor::render()
+	 * @inheritdoc
 	 */
 	public function render($content)
 	{
